fix(spec): validate resume dependency coverage
